--- ReactApp/car_rental_react/src/routes/AuthPage.js
@@ -1,113 +1,0 @@
-import authStyling from "../css/auth.css";
-import { useLocation, Link } from "react-router-dom";
-import { useEffect, useState } from "react";
-
-import React from "react";
-
-function AuthPage() {
-  let [authMode, setAuthMode] = useState("signin");
-
-  const changeAuthMode = () => {
-    setAuthMode(authMode === "signin" ? "signup" : "signin");
-  };
-
-  if (authMode === "signin") {
-    return (
-      <section className="Auth-Section">
-        <div className="Auth-Page-Div">
-          <div className="Auth-form-container">
-            <form className="Auth-form">
-              <div className="Auth-form-content">
-                <h3 className="Auth-form-title">Member Sign In</h3>
-                <div className="text-center">
-                  Not registered yet?{" "}
-                  <span className="link-primary" onClick={changeAuthMode}>
-                    Sign Up
-                  </span>
-                </div>
-                <div className="form-group mt-3">
-                  <label className="authLabel">Email address</label>
-                  <input
-                    type="email"
-                    className="form-control mt-1"
-                    placeholder="Enter email"
-                  />
-                </div>
-                <div className="form-group mt-3">
-                  <label className="authLabel">Password</label>
-                  <input
-                    type="password"
-                    className="form-control mt-1"
-                    placeholder="Enter password"
-                  />
-                </div>
-                <div className="d-grid gap-2 mt-3">
-                  <button type="submit" className="btn btn-primary">
-                    Sign In
-                  </button>
-                </div>
-                <p className="text-center mt-2">
-                  Forgot <a href="#">password?</a>
-                </p>
-              </div>
-            </form>
-          </div>
-        </div>
-      </section>
-    );
-  }
-
-  return (
-    <section className="Auth-Section">
-      <div className="Auth-Page-Div">
-        <div className="Auth-form-container">
-          <form className="Auth-form">
-            <div className="Auth-form-content">
-              <h3 className="Auth-form-title">Sign Up</h3>
-              <div className="text-center">
-                Already registered?{" "}
-                <span className="link-primary" onClick={changeAuthMode}>
-                  Sign In
-                </span>
-              </div>
-              <div className="form-group mt-3">
-                <label className="authLabel">Full Name</label>
-                <input
-                  type="email"
-                  className="form-control mt-1"
-                  placeholder="e.g Jane Doe"
-                />
-              </div>
-              <div className="form-group mt-3">
-                <label className="authLabel">Email address</label>
-                <input
-                  type="email"
-                  className="form-control mt-1"
-                  placeholder="Email Address"
-                />
-              </div>
-              <div className="form-group mt-3">
-                <label className="authLabel">Password</label>
-                <input
-                  type="password"
-                  className="form-control mt-1"
-                  placeholder="Password"
-                />
-              </div>
-              <div className="d-grid gap-2 mt-3">
-                <button type="submit" className="btn btn-primary">
-                  Create Account
-                </button>
-              </div>
-              <p className="text-center mt-2">
-                Forgot <a href="#">password?</a>
-              </p>
-            </div>
-          </form>
-        </div>
-      </div>
-    </section>
-  );
-}
-
-export default AuthPage;
